+++ Kernel/System/Calendar/Appointment.pm
@@ -1,5 +1,6 @@
 # --
 # Copyright (C) 2001-2021 OTRS AG, https://otrs.com/
+# Copyright (C) 2021 Znuny GmbH, https://znuny.org/
 # --
 # This software comes with ABSOLUTELY NO WARRANTY. For details, see
 # the enclosed file COPYING for license information (GPL). If you
@@ -2574,7 +2575,7 @@ sub _CalculateRecurrenceTime {
         my $Found;
 
         # loop up to 7*n times (7 days in week * frequency)
-        LOOP:
+        DAY:
         for ( my $Counter = 0; $Counter < 7 * $Param{Appointment}->{RecurrenceInterval}; $Counter++ ) {
 
             # Add one day.
@@ -2587,14 +2588,14 @@ sub _CalculateRecurrenceTime {
                 OriginalTime => $OriginalTimeObject,
             );
 
-            next LOOP if $CWDiff % $Param{Appointment}->{RecurrenceInterval};
+            next DAY if $CWDiff % $Param{Appointment}->{RecurrenceInterval};
 
             my $WeekDay = $CurrentTimeObject->Get()->{DayOfWeek};
 
             # check if SystemTime match requirements
             if ( grep { $WeekDay == $_ } @{ $Param{Appointment}->{RecurrenceFrequency} } ) {
                 $Found = 1;
-                last LOOP;
+                last DAY;
             }
         }