+++ Kernel/System/Console/BaseCommand.pm
@@ -1,5 +1,6 @@
 # --
 # Copyright (C) 2001-2021 OTRS AG, https://otrs.com/
+# Copyright (C) 2021 Znuny GmbH, https://znuny.org/
 # --
 # This software comes with ABSOLUTELY NO WARRANTY. For details, see
 # the enclosed file COPYING for license information (GPL). If you
@@ -697,7 +698,7 @@ sub TableOutput {
         }
     }
 
-    Row:
+    ROW:
     for my $Row ( @{ $Param{TableData}->{Body} } ) {
 
         next ROW if !$Row;