+++ Kernel/System/CustomerUser.pm
@@ -1,5 +1,6 @@
 # --
 # Copyright (C) 2001-2021 OTRS AG, https://otrs.com/
+# Copyright (C) 2021 Znuny GmbH, https://znuny.org/
 # --
 # This software comes with ABSOLUTELY NO WARRANTY. For details, see
 # the enclosed file COPYING for license information (GPL). If you
@@ -546,7 +547,7 @@ sub CustomerUserSearchFields {
                 Source    => $Param{Source},     # to get the right database field for the given source
             );
 
-            next SEARCHFIELDNAME if !%FieldConfig;
+            next ENTRY if !%FieldConfig;
 
             my %SearchFieldData = (
                 %FieldConfig,